+++ lib/iris/tests/unit/fileformats/cf/test_CFReader.py
@@ -70,7 +70,10 @@ class Test_translate__global_attributes(tests.IrisTest):
         )
 
     def test_create_global_attributes(self):
-        with mock.patch("netCDF4.Dataset", return_value=self.dataset):
+        with mock.patch(
+            "iris.fileformats.netcdf._thread_safe_nc.DatasetWrapper",
+            return_value=self.dataset,
+        ):
             global_attrs = CFReader("dummy").cf_group.global_attributes
             self.assertEqual(
                 global_attrs["dimensions"], "something something_else"
@@ -145,7 +148,10 @@ class Test_translate__formula_terms(tests.IrisTest):
         self.addCleanup(reset_patch.stop)
 
     def test_create_formula_terms(self):
-        with mock.patch("netCDF4.Dataset", return_value=self.dataset):
+        with mock.patch(
+            "iris.fileformats.netcdf._thread_safe_nc.DatasetWrapper",
+            return_value=self.dataset,
+        ):
             cf_group = CFReader("dummy").cf_group
             self.assertEqual(len(cf_group), len(self.variables))
             # Check there is a singular data variable.
@@ -247,7 +253,10 @@ class Test_build_cf_groups__formula_terms(tests.IrisTest):
         self.addCleanup(patcher.stop)
 
     def test_associate_formula_terms_with_data_variable(self):
-        with mock.patch("netCDF4.Dataset", return_value=self.dataset):
+        with mock.patch(
+            "iris.fileformats.netcdf._thread_safe_nc.DatasetWrapper",
+            return_value=self.dataset,
+        ):
             cf_group = CFReader("dummy").cf_group
             self.assertEqual(len(cf_group), len(self.variables))
             # Check the cf-group associated with the data variable.
@@ -296,7 +305,10 @@ class Test_build_cf_groups__formula_terms(tests.IrisTest):
                 )
 
     def test_promote_reference(self):
-        with mock.patch("netCDF4.Dataset", return_value=self.dataset):
+        with mock.patch(
+            "iris.fileformats.netcdf._thread_safe_nc.DatasetWrapper",
+            return_value=self.dataset,
+        ):
             cf_group = CFReader("dummy").cf_group
             self.assertEqual(len(cf_group), len(self.variables))
             # Check the number of data variables.
@@ -316,7 +328,8 @@ class Test_build_cf_groups__formula_terms(tests.IrisTest):
     def test_formula_terms_ignore(self):
         self.orography.dimensions = ["lat", "wibble"]
         with mock.patch(
-            "netCDF4.Dataset", return_value=self.dataset
+            "iris.fileformats.netcdf._thread_safe_nc.DatasetWrapper",
+            return_value=self.dataset,
         ), mock.patch("warnings.warn") as warn:
             cf_group = CFReader("dummy").cf_group
             group = cf_group.promoted
@@ -327,7 +340,8 @@ class Test_build_cf_groups__formula_terms(tests.IrisTest):
     def test_auxiliary_ignore(self):
         self.x.dimensions = ["lat", "wibble"]
         with mock.patch(
-            "netCDF4.Dataset", return_value=self.dataset
+            "iris.fileformats.netcdf._thread_safe_nc.DatasetWrapper",
+            return_value=self.dataset,
         ), mock.patch("warnings.warn") as warn:
             cf_group = CFReader("dummy").cf_group
             promoted = ["x", "orography"]
@@ -342,7 +356,8 @@ class Test_build_cf_groups__formula_terms(tests.IrisTest):
         self.variables["wibble"] = self.wibble
         self.orography.coordinates = "wibble"
         with mock.patch(
-            "netCDF4.Dataset", return_value=self.dataset
+            "iris.fileformats.netcdf._thread_safe_nc.DatasetWrapper",
+            return_value=self.dataset,
         ), mock.patch("warnings.warn") as warn:
             cf_group = CFReader("dummy").cf_group.promoted
             promoted = ["wibble", "orography"]
