+++ setup.cfg
@@ -52,7 +52,7 @@ install_requires =
     cftime>=1.5.0
     dask[array]>=2.26
     matplotlib>=3.5
-    netcdf4<1.6.1
+    netcdf4
     numpy>=1.19
     scipy
     shapely!=1.8.3
